builds, lint, and tests are now all passing! :)
