+++ packages/common/.eslintignore
@@ -1,0 +1,2 @@
+dist/
+__tests__/fixtures/
