+++ packages/common/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "@kocal/eslint-config-generator",
-  "description": "An opinionated ESLint config generator for my projects. Comes with AirBnB, Vue, TypeScript and Prettier support.",
+  "name": "@kocal/eslint-config-generator-common",
+  "description": "Common things for @kocal/eslint-config-generator and plugins.",
   "publishConfig": {
     "access": "public"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/kocal/eslint-config-generator.git",
-    "directory": "packages/eslint-config-generator"
+    "directory": "packages/common"
   },
   "author": "Hugo Alliaume <hugo@alliau.me>",
   "license": "MIT",
@@ -33,27 +33,15 @@
     "README.md",
     "LICENSE"
   ],
-  "dependencies": {
-    "@kocal/eslint-config-generator-common": "workspace:*",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-config-prettier": "^6.13.0",
-    "eslint-import-resolver-node": "^0.3.4",
-    "eslint-import-resolver-webpack": "^0.13.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.1.4"
-  },
   "peerDependencies": {
-    "@babel/eslint-parser": "7.0.0",
-    "eslint": "^7.0.0",
-    "prettier": "^2.0.0"
+    "typescript": "^4.0.0",
+    "eslint": "^7.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.3",
     "@babel/eslint-parser": "^7.12.1",
     "@babel/preset-env": "^7.12.1",
     "@babel/preset-typescript": "^7.12.1",
-    "@kocal/eslint-config-generator-plugin-typescript": "workspace:*",
-    "@kocal/eslint-config-generator-plugin-vue": "workspace:*",
     "@rollup/plugin-babel": "^5.2.1",
     "@rollup/plugin-node-resolve": "^9.0.0",
     "@types/eslint": "^7.2.4",
@@ -64,9 +52,6 @@
     "jest-standard-reporter": "^2.0.0",
     "prettier": "^2.1.2",
     "rollup": "^2.32.0",
-    "typescript": "^4.0.3",
-    "vue": "^2.6.12",
-    "vue-class-component": "^7.2.6",
-    "vue-property-decorator": "^9.1.2"
+    "typescript": "^4.0.3"
   }
 }
