+++ packages/common/src/options.ts
