+++ packages/common/tsconfig.json
@@ -6,5 +6,8 @@
   },
   "include": [
     "src/**/*"
+  ],
+  "exclude": [
+    "src/**/*.test.ts"
   ]
 }
