+++ packages/eslint-config-generator/src/index.ts
@@ -1,4 +1,1 @@
-export * from './config';
 export * from './config-generator';
-export * from './options';
-export * from './plugin';
