+++ packages/eslint-config-generator/tsconfig.json
@@ -3,7 +3,6 @@
   "compilerOptions": {
     "declaration": true,
     "declarationDir": "dist",
-    "types": ["@types/jest", "@types/eslint", "@types/node"]
   },
   "include": [
     "src/**/*"
