--- packages/plugin-vue/__tests__/fixtures/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-  "root": true
-}
