--- packages/plugin-vue/__tests__/functional.test.ts
@@ -1,524 +1,0 @@
-import { ESLint } from 'eslint';
-import * as path from 'path';
-import { generateConfig } from '@kocal/eslint-config-generator';
-
-let eslint: ESLint;
-
-function createESLint(options: ESLint.Options): ESLint {
-  return new ESLint({
-    cwd: path.join(__dirname, 'fixtures'),
-    ignore: false,
-    ...options,
-  });
-}
-
-function cleanResults(results: ESLint.LintResult[]) {
-  return results.map((result) => {
-    return {
-      ...result,
-      messages: result.messages.map(({ line, column, ruleId, message }) => {
-        return { line, column, ruleId, message };
-      }),
-    };
-  });
-}
-
-async function lintCode(
-  code: string,
-  options?: { filePath?: string; warnIgnored?: boolean }
-): Promise<ESLint.LintResult[]> {
-  return eslint.lintText(code, options);
-}
-
-async function lintVueCode(code: string): Promise<ESLint.LintResult[]> {
-  return lintCode(code, { filePath: 'file.vue' });
-}
-
-describe('Functional', function () {
-  it('should lint Vue code', async function () {
-    eslint = createESLint({
-      baseConfig: generateConfig({
-        vue: true,
-      }),
-    });
-
-    const results = cleanResults(await eslint.lintFiles('component.vue'));
-
-    expect(results[0].messages).toMatchInlineSnapshot(`
-      Array [
-        Object {
-          "column": 5,
-          "line": 4,
-          "message": "Component name \\"transition\\" is not PascalCase.",
-          "ruleId": "vue/component-name-in-template-casing",
-        },
-        Object {
-          "column": 29,
-          "line": 15,
-          "message": "Delete \`·\`",
-          "ruleId": "prettier/prettier",
-        },
-        Object {
-          "column": 5,
-          "line": 20,
-          "message": "Component name \\"v-component\\" is not PascalCase.",
-          "ruleId": "vue/component-name-in-template-casing",
-        },
-        Object {
-          "column": 5,
-          "line": 32,
-          "message": "Unexpected console statement.",
-          "ruleId": "no-console",
-        },
-        Object {
-          "column": 17,
-          "line": 32,
-          "message": "Replace \`\\"Hello·world!\\"\` with \`'Hello·world!'\`",
-          "ruleId": "prettier/prettier",
-        },
-      ]
-    `);
-  });
-
-  it('should lint Vue code with TypeScript support on .vue files', async function () {
-    eslint = createESLint({
-      baseConfig: generateConfig({
-        vue: true,
-        typescript: {
-          vueComponents: true,
-        },
-      }),
-    });
-
-    const results = cleanResults(await eslint.lintFiles('component-typescript.vue'));
-
-    expect(results[0].messages).toMatchInlineSnapshot(`
-      Array [
-        Object {
-          "column": 5,
-          "line": 4,
-          "message": "Component name \\"transition\\" is not PascalCase.",
-          "ruleId": "vue/component-name-in-template-casing",
-        },
-        Object {
-          "column": 29,
-          "line": 5,
-          "message": "Replace \`⏎········my·image:·<img·src=\\"https://example.com/image.png\\"·/>⏎······\` with \`my·image:·<img·src=\\"https://example.com/image.png\\"·/>\`",
-          "ruleId": "prettier/prettier",
-        },
-        Object {
-          "column": 28,
-          "line": 13,
-          "message": "Delete \`·\`",
-          "ruleId": "prettier/prettier",
-        },
-        Object {
-          "column": 5,
-          "line": 18,
-          "message": "Component name \\"v-component\\" is not PascalCase.",
-          "ruleId": "vue/component-name-in-template-casing",
-        },
-        Object {
-          "column": 3,
-          "line": 27,
-          "message": "Type boolean trivially inferred from a boolean literal, remove type annotation.",
-          "ruleId": "@typescript-eslint/no-inferrable-types",
-        },
-        Object {
-          "column": 3,
-          "line": 29,
-          "message": "Missing return type on function.",
-          "ruleId": "@typescript-eslint/explicit-module-boundary-types",
-        },
-        Object {
-          "column": 10,
-          "line": 29,
-          "message": "Expected 'this' to be used by class method 'mounted'.",
-          "ruleId": "class-methods-use-this",
-        },
-        Object {
-          "column": 5,
-          "line": 30,
-          "message": "Unexpected console statement.",
-          "ruleId": "no-console",
-        },
-        Object {
-          "column": 17,
-          "line": 30,
-          "message": "Replace \`\\"Hello·world!\\"\` with \`'Hello·world!'\`",
-          "ruleId": "prettier/prettier",
-        },
-        Object {
-          "column": 12,
-          "line": 34,
-          "message": "Expected 'this' to be used by class method 'onVisible'.",
-          "ruleId": "class-methods-use-this",
-        },
-        Object {
-          "column": 5,
-          "line": 35,
-          "message": "Unexpected console statement.",
-          "ruleId": "no-console",
-        },
-      ]
-    `);
-  });
-
-  describe('Rule', function () {
-    describe('vue/no-bare-strings-in-template', function () {
-      it('should fail', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>Hello</div>
-</template>
-`);
-
-        expect(results[0].messages).toMatchInlineSnapshot(`Array []`);
-      });
-
-      it('should pass', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>
-    {{ $t('hello') }}
-  </div>
-</template>
-`);
-
-        expect(results[0].messages).toEqual([]);
-      });
-    });
-
-    describe('vue/no-duplicate-attr-inheritance', function () {
-      it('should fail', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>
-    <MyComponent v-bind="$attrs" />
-  </div>
-</template>
-`);
-
-        expect(results[0].messages).toMatchInlineSnapshot(`
-          Array [
-            Object {
-              "column": 26,
-              "endColumn": 32,
-              "endLine": 3,
-              "line": 3,
-              "message": "Set \\"inheritAttrs\\" to false.",
-              "nodeType": "Identifier",
-              "ruleId": "vue/no-duplicate-attr-inheritance",
-              "severity": 2,
-            },
-          ]
-        `);
-      });
-
-      it('should pass', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>
-    <MyComponent v-bind="$attrs" />
-  </div>
-</template>
-
-<script>
-export default {
-  inheritAttrs: false,
-};
-</script>
-`);
-
-        expect(results[0].messages).toEqual([]);
-      });
-    });
-
-    describe('vue/no-empty-component-block', function () {
-      it('should fail', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<style></style>
-`);
-
-        expect(results[0].messages).toMatchInlineSnapshot(`
-          Array [
-            Object {
-              "column": 1,
-              "endColumn": 16,
-              "endLine": 1,
-              "line": 1,
-              "message": "\`<style>\` is empty. Empty block is not allowed.",
-              "messageId": "unexpected",
-              "nodeType": "VElement",
-              "ruleId": "vue/no-empty-component-block",
-              "severity": 2,
-            },
-          ]
-        `);
-      });
-
-      it('should pass', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>
-    <MyComponent />
-  </div>
-</template>
-`);
-
-        expect(results[0].messages).toEqual([]);
-      });
-    });
-
-    describe('vue/no-template-target-blank', function () {
-      it('should fail', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <a href="https://example.com" target="_blank">{{ $t('example') }}</a>
-</template>
-`);
-
-        expect(results[0].messages).toMatchInlineSnapshot(`
-          Array [
-            Object {
-              "column": 33,
-              "endColumn": 48,
-              "endLine": 2,
-              "line": 2,
-              "message": "Using target=\\"_blank\\" without rel=\\"noopener noreferrer\\" is a security risk.",
-              "nodeType": "VAttribute",
-              "ruleId": "vue/no-template-target-blank",
-              "severity": 2,
-            },
-          ]
-        `);
-      });
-
-      it('should pass', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <a href="https://example.com" target="_blank" rel="noopener noreferrer">{{ $t('example') }}</a>
-</template>
-`);
-
-        expect(results[0].messages).toEqual([]);
-      });
-    });
-
-    describe('vue/padding-line-between-blocks', function () {
-      it('should fail', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>Hello world</div>
-</template>
-<script>
-export default {};
-</script>
-`);
-
-        expect(results[0].messages).toMatchInlineSnapshot(`
-            Array [
-              Object {
-                "column": 1,
-                "endColumn": 10,
-                "endLine": 6,
-                "fix": Object {
-                  "range": Array [
-                    47,
-                    47,
-                  ],
-                  "text": "
-            ",
-                },
-                "line": 4,
-                "message": "Expected blank line before this block.",
-                "messageId": "always",
-                "nodeType": "VElement",
-                "ruleId": "vue/padding-line-between-blocks",
-                "severity": 2,
-              },
-            ]
-          `);
-      });
-
-      it('should pass', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>Hello world</div>
-</template>
-
-<script>
-export default {};
-</script>
-`);
-
-        expect(results[0].messages).toEqual([]);
-      });
-    });
-
-    describe('vue/v-on-function-call', function () {
-      it('should fail', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>
-    <button @click="closeModal()">Button</button>
-  </div>
-</template>
-`);
-
-        expect(results[0].messages).toMatchInlineSnapshot(`
-          Array [
-            Object {
-              "column": 21,
-              "endColumn": 33,
-              "endLine": 3,
-              "fix": Object {
-                "range": Array [
-                  39,
-                  51,
-                ],
-                "text": "closeModal",
-              },
-              "line": 3,
-              "message": "Method calls without arguments inside of 'v-on' directives must not have parentheses.",
-              "nodeType": "CallExpression",
-              "ruleId": "vue/v-on-function-call",
-              "severity": 2,
-            },
-          ]
-        `);
-      });
-
-      it('should pass', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<template>
-  <div>
-    <button @click="closeModal">Button</button>
-    <button @click="closeModal(args)">Button</button>
-  </div>
-</template>
-`);
-
-        expect(results[0].messages).toEqual([]);
-      });
-    });
-
-    describe('vue/no-boolean-default', function () {
-      it('should fail', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<script>
-export default {
-  props: {
-    rounded: {
-      type: Boolean,
-      default: true,
-    },
-  },
-};
-</script>
-`);
-
-        expect(results[0].messages).toMatchInlineSnapshot(`
-          Array [
-            Object {
-              "column": 7,
-              "endColumn": 20,
-              "endLine": 6,
-              "line": 6,
-              "message": "Boolean prop should not set a default (Vue defaults it to false).",
-              "nodeType": "Property",
-              "ruleId": "vue/no-boolean-default",
-              "severity": 2,
-            },
-          ]
-        `);
-      });
-
-      it('should pass', async function () {
-        eslint = createESLint({
-          baseConfig: generateConfig({
-            vue: true,
-          }),
-        });
-
-        const results = await lintVueCode(`<script>
-export default {
-  props: {
-    rounded: {
-      type: Boolean,
-    },
-  },
-};
-</script>
-`);
-
-        expect(results[0].messages).toEqual([]);
-      });
-    });
-  });
-});
