+++ packages/plugin-vue/package.json
@@ -18,7 +18,7 @@
     "rebuild": "npm run build",
     "test": "npm run test:ts && npm run test:unit",
     "test:ts": "tsc --noEmit",
-    "test:unit": "jest",
+    "test:unit": "echo 'No tests.'",
     "lint": "eslint . --ext .js,.ts",
     "lint:fix": "npm run lint -- --fix"
   },
@@ -34,6 +34,7 @@
     "LICENSE"
   ],
   "dependencies": {
+    "@kocal/eslint-config-generator-common": "workspace:*",
     "eslint-plugin-vue": "^7.1.0"
   },
   "peerDependencies": {
@@ -45,20 +46,12 @@
     "@babel/eslint-parser": "^7.12.1",
     "@babel/preset-env": "^7.12.1",
     "@babel/preset-typescript": "^7.12.1",
-    "@kocal/eslint-config-generator": "workspace:*",
-    "@kocal/eslint-config-generator-plugin-typescript": "workspace:*",
     "@rollup/plugin-babel": "^5.2.1",
     "@rollup/plugin-node-resolve": "^9.0.0",
     "@types/eslint": "^7.2.4",
-    "@types/jest": "^26.0.14",
     "eslint": "^7.11.0",
-    "jest": "^26.6.0",
-    "jest-standard-reporter": "^2.0.0",
     "prettier": "^2.1.2",
     "rollup": "^2.32.0",
-    "typescript": "^4.0.3",
-    "vue-class-component": "^7.2.6",
-    "vue-property-decorator": "^9.0.2",
-    "vue": "^2.6.12"
+    "typescript": "^4.0.3"
   }
 }
