+++ packages/plugin-vue/tsconfig.json
@@ -6,8 +6,5 @@
   },
   "include": [
     "src/**/*"
-  ],
-  "exclude": [
-    "src/**/*.test.ts"
   ]
 }
